requirements.txt
apify-client
bertopic
flask
python-dotenv
sentence-transformers
scikit-learn
umap-learn
hdbscan
pandas
numpy
